backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
motor==3.4.0
redis==5.0.4
aioredis==2.0.1
cloudinary==1.39.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python-headless==4.9.0.80
scikit-image==0.23.1
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
timm==0.9.16
transformers==4.40.1
xgboost==2.0.3
shap==0.45.1
librosa==0.10.1
torch==2.2.2
torchvision==0.17.2
python-telegram-bot[job-queue]==21.1.1
reportlab==4.1.0
httpx==0.27.0
beautifulsoup4==4.12.3
loguru==0.7.2
slowapi==0.1.9
python-magic==0.4.27
anthropic==0.25.0
python-multipart==0.0.9
asyncpg==0.29.0
aiofiles==23.2.1\n